requirements.txt
pyserial
vispy
numpy
PyQt5